Added Student and Professor Roles
Added roles for Students and Professors. If a endpoint should only be accessible to a professor or a student use the @Secured(PROFESSOR_ROLE) or @Secured(STUDENT_ROLE) annotation above the controller method. Also I added logic so that users that are logged in will be redirected to the homepage if they try to make a GET request to the login or register pages.